File Split and Adaptive dt (BGS + MVJ)
