+++ cupydo/algorithm/__init__.py
@@ -1,0 +1,3 @@
+from .BGS import *
+from .IQN_ILS import *
+from .IQN_MVJ import *
