+++ cupydo/interfaces/ExampleSolid.py
@@ -108,7 +108,7 @@ class ExampSolver(SolidSolver):
     
         #return no
 
-    def applyNodalLoads(self, load_X, load_Y, load_Z, val_time, haloNodesLoads = {}):
+    def applyNodalLoads(self, load_X, load_Y, load_Z, dt, haloNodesLoads = {}):
         """
         Des.
         """
@@ -119,7 +119,7 @@ class ExampSolver(SolidSolver):
             #self.coreSolver.applyNodalLoadY(load_Y[ii], ii)
             #self.coreSolver.applyNodalLoadZ(load_Z[ii], ii)
 
-    def applyNodalTemperatures(self, Temperature, val_time):
+    def applyNodalTemperatures(self, Temperature, dt):
         """
         Des.
         """
