+++ cupydo/interfaces/GetDP.py
@@ -374,7 +374,7 @@ class GetDP(SolidSolver):
 
         return self.nodalInterfIndex[iVertex]
 
-    def applyNodalLoads(self, load_X, load_Y, load_Z, time, haloNodesLoads = {}):
+    def applyNodalLoads(self, load_X, load_Y, load_Z, dt, haloNodesLoads = {}):
         """
         Des.
         """
@@ -385,7 +385,7 @@ class GetDP(SolidSolver):
             self.__writeVecToFile("nodalForce.txt", load_X, load_Y, load_Z, self.nodalInterfIndex)
 
 
-    def applyNodalTemperatures(self, Temperature, val_time):
+    def applyNodalTemperatures(self, Temperature, dt):
         """
         Des.
         """
@@ -395,7 +395,7 @@ class GetDP(SolidSolver):
         else:
             self.__writeScalToFile("nodalTemperature.txt" , Temperature, self.nodalInterfIndex)
 
-    def applyNodalNormalHeatFluxes(self, NormalHeatFlux, val_time):
+    def applyNodalNormalHeatFluxes(self, NormalHeatFlux, dt):
         """
         Des.
         """
@@ -405,7 +405,7 @@ class GetDP(SolidSolver):
         else:
             self.__writeScalToFile("nodalNormalHeatFlux.txt", NormalHeatFlux, self.nodalInterfIndex)
 
-    def applyNodalHeatFluxes(self, HeatFlux_X, HeatFlux_Y, HeatFlux_Z, time):
+    def applyNodalHeatFluxes(self, HeatFlux_X, HeatFlux_Y, HeatFlux_Z, dt):
         """
         Des.
         """
