+++ cupydo/interfaces/Pfem3D.py
@@ -1,3 +1,28 @@
+#! /usr/bin/env python3
+# -*- coding: utf8 -*-
+
+''' 
+
+Copyright 2018 University of Liège
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License. 
+
+Metafor.py
+Python interface between the wrapper of Metafor and CUPyDO.
+Authors R. BOMAN, M.L. CERQUAGLIA, D. THOMAS
+
+'''
+
 from ..genericSolvers import FluidSolver
 from ..utilities import titlePrint
 import pfem3Dw as w
@@ -79,8 +104,7 @@ class Pfem3D(FluidSolver):
                 
                 dt = float(dt/2)
                 count = np.multiply(2,count)
-                if dt < (t2-t1)/self.maxDivision:
-                    raise Exception('Too large time step')
+                if dt < (t2-t1)/self.maxDivision: return False
                 continue
 
             count = count-1
@@ -99,8 +123,7 @@ class Pfem3D(FluidSolver):
 
         self.solver.computeNextDT()
         division = int((t2-t1)/self.solver.getTimeStep())
-        if division > self.maxDivision:
-            raise Exception('Too large time step')
+        if division > self.maxDivision: return False
         dt = (t2-t1)/division
 
         # Main solving loop for the fluid simulation
@@ -116,10 +139,10 @@ class Pfem3D(FluidSolver):
 
 # %% Apply Boundary Conditions
 
-    def applyNodalDisplacements(self,dx,dy,dz,*_):
+    def applyNodalDisplacements(self,dx,dy,dz,dx_nM1,dy_nM1,dz_nM1,haloNodesDisplacements,dt):
 
-        BC = (np.transpose([dx,dy,dz])-self.disp)/self.dt
-        if not self.implicit: BC = 2*(BC-self.vel)/self.dt
+        BC = (np.transpose([dx,dy,dz])-self.disp)/dt
+        if not self.implicit: BC = 2*(BC-self.vel)/dt
 
         for i,vector in enumerate(BC):
             for j,val in enumerate(vector): self.BC[i][j] = val
@@ -163,7 +186,7 @@ class Pfem3D(FluidSolver):
 
 # %% Other Functions
 
-    def update(self,_):
+    def update(self,dt):
 
         self.mesh.remesh(False)
         if self.implicit: self.solver.precomputeMatrix()
