+++ cupydo/interpolator/__init__.py
@@ -1,0 +1,3 @@
+from .conservative import *
+from .consistent import *
+from .matching import *
